requirements.txt
asgiref
attrs
boto3
botocore
CacheControl
cachetools
certifi
cffi
chardet
charset-normalizer
click
coreapi
coreschema
cryptography
defusedxml
diff-match-patch
Django
django-ckeditor
django-cors-headers
django-filter
django-import-export
django-js-asset
django-modeltranslation
django-rest-swagger
django-storages
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-yasg
et-xmlfile
firebase-admin
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
googletrans
grpcio
grpcio-status
gTTS
h11
h2
hpack
hstspreload
httpcore
httplib2
httpx
hyperframe
idna
inflection
itypes
Jinja2
jmespath
jsonschema
jsonschema-specifications
Markdown
MarkupPy
MarkupSafe
msgpack
numpy
odfpy
openapi-codec
opencv-python
openpyxl
packaging
Pillow
polib
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyparsing
pypng
python-barcode
python-dateutil
